Move all file-related classes and variables into dedicated file
